Moving service state tests to pytest.
