+++ tests/filename_map.yml
@@ -23,7 +23,7 @@ salt/modules/(aix_group|groupadd|mac_group|pw_group|solaris_group|win_groupadd)\
   - unit.states.test_group
 
 salt/modules/(debian_service|freebsdservice|gentoo_service|launchctl_service|mac_service|netbsdservice|openbsdrcctl_service|openbsdservice|rh_service|runit|linux_service|smf_service|systemd_service|upstart_service|win_service)\.py:
-  - unit.states.test_service
+  - pytests.unit.states.test_service
   - integration.modules.test_service
   - integration.states.test_service
